added error for integer overflow
For factorials with large n the integers were too large for the int64 and would
overflow. Added detection for this by checking if n is negative after each
level of recursion in the factorial function (not a great way to check for
overflow, it's a bit hacky). To add support for binomial probability with large
n (~20) an arbitrary precision number library / system would need to be put in
place.